Moved integration tests from UserControllerTest.kt into UserControllerIntegrationTest.kt
