+++ backend/src/test/kotlin/za/ac/up/cs/atbash/controller/UserControllerTest.kt
@@ -24,22 +24,6 @@ class UserControllerTest {
     //------loginTestCases------//
 
     @Test
-    @DisplayName("When LoginRequest 'number' is null, response status should be BAD_REQUEST")
-    fun loginWhenRequestNumberNull(){
-        val response = userController.login(LoginRequestJson(null, "password"))
-
-        Assertions.assertEquals(HttpStatus.BAD_REQUEST, response.statusCode)
-    }
-
-    @Test
-    @DisplayName("When LoginRequest 'password' is null, response status should be BAD_REQUEST")
-    fun loginWhenRequestPasswordNull(){
-        val response = userController.login(LoginRequestJson("number", null))
-
-        Assertions.assertEquals(HttpStatus.BAD_REQUEST, response.statusCode)
-    }
-
-    @Test
     @DisplayName("When user requesting to login does not exist, response status should be INTERNAL_SERVER_ERROR")
     fun loginWhenServiceUnsuccessful(){
         Mockito.`when`(userService.verifyLogin(Mockito.anyString(), Mockito.anyString())).thenReturn(null)
@@ -58,30 +42,6 @@ class UserControllerTest {
     }
 
     //------registerUseCases------//
-
-    @Test
-    @DisplayName("When RegisterRequest 'number' is null, response status should be BAD_REQUEST")
-    fun registerWhenRequestNumberNull(){
-        val response = userController.register(RegisterRequestJson(null, "password", "deviceToken"))
-
-        Assertions.assertEquals(HttpStatus.BAD_REQUEST, response.statusCode)
-    }
-
-    @Test
-    @DisplayName("When RegisterRequest 'password' is null, response status should be BAD_REQUEST")
-    fun registerWhenRequestPasswordNull(){
-        val response = userController.register(RegisterRequestJson("number", null, "deviceToken"))
-
-        Assertions.assertEquals(HttpStatus.BAD_REQUEST, response.statusCode)
-    }
-
-    @Test
-    @DisplayName("When RegisterRequest 'deviceToken' is null, response status should be BAD_REQUEST")
-    fun registerWhenRequestDeviceTokenNull(){
-        val response = userController.register(RegisterRequestJson("number", "password", null))
-
-        Assertions.assertEquals(HttpStatus.BAD_REQUEST, response.statusCode)
-    }
 
     @Test
     @DisplayName("When user requesting to register already exists, response status should be INTERNAL_SERVER_ERROR")
